tests/wpt/web-platform-tests/shadow-dom/user-interaction/ranges-and-selections/test-001.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_07_01_01</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#ranges-and-selection">
<meta name="assert" content="User Interaction: Selection, returned by the window.getSelection() method must never return a selection within a shadow tree">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(unit(function (ctx) {
	var d = newRenderedHTMLDocument(ctx);

	var host = d.createElement('div');
	d.body.appendChild(host);
	var s = host.createShadowRoot();

	var span = d.createElement('span');
	span.innerHTML = 'Some text';
	s.appendChild(span);

	var range = d.createRange();
	range.setStart(span.firstChild, 0);
	range.setEnd(span.firstChild, 3);

	var selection = window.getSelection();
    selection.removeAllRanges();
    selection.addRange(range);

    var sl = window.getSelection();
    assert_equals(sl.toString(), '', 'window.getSelection() method must never return a selection ' +
    		'within a shadow tree');

}), 'A_07_07_01_T01');


// test distributed nodes
test(unit(function (ctx) {
	var d = newRenderedHTMLDocument(ctx);

	var host = d.createElement('div');
	d.body.appendChild(host);

	var span = d.createElement('span');
	span.innerHTML = 'Some text';
	host.appendChild(span);

	var s = host.createShadowRoot();
	s.innerHTML = '<content select="span"></content>';

	var range = d.createRange();
	range.setStart(span.firstChild, 0);
	range.setEnd(span.firstChild, 3);

	var selection = window.getSelection();
    selection.removeAllRanges();
    selection.addRange(range);

    var sl = window.getSelection();
    assert_equals(sl.toString(), '', 'window.getSelection() method must never return a selection ' +
    	'within a shadow tree');

}), 'A_07_07_01_T02');
</script>
</body>
</html>
